services/auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
